Lots of reworking and the addition of the testing project
Rather than define romanisation patterns in an XML file, the rules of a
particular romanisation are now defined programmatically in the class
responsible for that system. This has the advantage that complex rules
can be expressed more coherently, and it also avoids having to parse
XML.

There is also a new test project, which is vital, because otherwise it
would be very difficult to verify that the code was working properly.