Addition of Jamo class for dealing with single characters, ability to romanise pronunciation changes, and more unit tests
The Jamo class has been added to represent single Korean characters,
which have different Unicode values. It works similarly to the Syllable
class. It's useful for removing magic strings and numbers from the rest
of the code.

The Jamo class has made it easier to define the rules for romanisation
of pronunciation changes, such as k + n become ngn. The rules for the
McCune-Reischauer romanisation class have been further completed, and
some unit tests have been written to evaluate them.