Integrating the Jamo and Syllable classes and adding more unit tests
Even though the Jamo and Syllable classes are similar, in that they
represent Unicode characters in specific ranges, and there's only an
implicit relation between those Unicode characters, I've made the two
classes work together so that the initial, medial, and final of a
syllable can be represented as the individual jamo, rather than just as
indices. This makes it easier to use jamo constants in the program and
remove magic numbers.

I've also added in more unit tests for two-syllable 'words', so as to
test romanisation changes due to changes in pronunciation. There are
some romanisation changes that I have yet to characterise in the code.